+++ src/components/passwordFind/PasswordFindForm.jsx
@@ -1,9 +1,11 @@
 // 비번찾기 페이지.
 import { useState } from 'react';
-import { useEffect } from 'react';
 import { handleEmailchecking, handleConfirmCodechecking } from '../../utils/register.js';
+import { inputChange } from '../../utils/inputOnChange.js';
 import styles from './PasswordFindForm.module.css';
-import { APIService } from '@api/axios';
+
+import { handleSendingClick, handleCheckingClick, useTimerEmailConfirm } from '../../hooks/useEmailConfirmHook.js';
+import { subPasswordGet } from '../../hooks/usePasswordFindHook.js';
 
 export default function PasswordFindForm({ emailSuccess, setEmailSuccess, setEmail, setNow, setSubPassword }) {
   const [form, setForm] = useState({
@@ -25,181 +27,7 @@ export default function PasswordFindForm({ emailSuccess, setEmailSuccess, setEma
   const [confirmSuccess, setConfirmSuccess] = useState(1);
   const [confirms, setConfirms] = useState({});
 
-  // 타이머 관련 함수 //
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setCount((count) => count - 1);
-    }, 1000);
-
-    if (count === 0) {
-      clearInterval(timer);
-    }
-
-    return () => clearInterval(timer);
-  }, [form.timing]);
-
-  useEffect(() => {
-    setM(Math.floor(count / 60));
-    setS(count % 60);
-
-    if (count === 0 || count < 0) {
-      setForm({ ...form, timing: false, confirmCode: '' });
-      setErrors({
-        ...form,
-        confirmCode: '입력 시간이 만료되었습니다. 다시 인증번호를 전송해주세요.',
-      });
-    }
-  }, [count]);
-
-  function inputChange(event) {
-    const { id, value } = event.target;
-    setForm({ ...form, [id]: value });
-
-    if (id === 'email') {
-      if (value === '') {
-        setSendSuccess(1);
-      } else {
-        setSendSuccess(2);
-      }
-    } else if (id === 'confirmCode') {
-      if (value === '') {
-        setConfirmSuccess(1);
-      } else {
-        setConfirmSuccess(2);
-      }
-    }
-  }
-
-  // 인증번호 전송 버튼 클릭 //
-  async function handleSendingClick(event) {
-    event.preventDefault();
-    const isValid = handleEmailchecking(setErrors, form);
-
-    if (isValid) {
-      try {
-        setSendSuccess(3);
-        // 이메일에 도메인을 붙여서 전송
-        const fullEmail = `${form.email}@skuniv.ac.kr`;
-        // 이메일 인증번호 전송 API 호출
-        const response = await APIService.public.post(import.meta.env.VITE_APP_AUTH_EMAIL_SEND, { email: fullEmail });
-
-        // 인증번호 이메일일 전송 성공시
-        if (response.success === true) {
-          setSendSuccess(2);
-          setConfirms({ ...form, email: '인증번호가 전송되었습니다.' });
-          setCount(300); // 5분
-          setForm({ ...form, email_valid: true, sendemail: form.email, timing: true });
-        } else {
-          setErrors({
-            ...errors,
-            email: '인증번호 전송에 실패했습니다.',
-          });
-        }
-      } catch (error) {
-        //에러처리
-        setErrors({
-          ...errors,
-          email: error.response?.data?.message || '인증번호 전송에 실패했습니다.',
-        });
-      }
-    }
-  }
-
-  // 인증번호 확인 버튼 클릭 //
-  async function handleCheckingClick(event) {
-    event.preventDefault();
-
-    const isValid = handleConfirmCodechecking(setErrors, form);
-    if (!isValid) return;
-
-    try {
-      const fullEmail = `${form.sendemail}@skuniv.ac.kr`;
-
-      const requestData = {
-        email: fullEmail,
-        code: String(form.confirmCode),
-      };
-
-      const response = await APIService.public.post(import.meta.env.VITE_APP_AUTH_EMAIL_VERIFY, requestData);
-
-      // verified가 false인 경우도 처리
-      if (response.verified === true) {
-        setConfirms((prev) => ({
-          ...prev,
-          confirmCode: response.message || '이메일이 인증되었습니다.',
-        }));
-
-        setEmailSuccess(true);
-        setForm((prev) => ({
-          ...prev,
-          confirmCode_valid: true,
-          timing: false,
-        }));
-      } else {
-        // 인증번호가 틀린 경우 (verified가 false인 경우)
-        setErrors((prev) => {
-          const newErrors = {
-            ...prev,
-            confirmCode: '잘못된 인증번호입니다. 다시 입력해주세요',
-          };
-          return newErrors;
-        });
-
-        // form의 confirmCode_valid 상태도 false로 설정
-        setForm((prev) => ({
-          ...prev,
-          confirmCode_valid: false,
-        }));
-      }
-    } catch {
-      // 서버 응답 자체가 실패한 경우
-      setErrors((prev) => ({
-        ...prev,
-        confirmCode: '서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요.',
-      }));
-
-      setForm((prev) => ({
-        ...prev,
-        confirmCode_valid: false,
-      }));
-    }
-  }
-
-  // 비밀번호 찾기 버튼 클릭 => 임시 비번 발급하고 성공하면 페이지 넘기기 //
-  function next(e) {
-    e.preventDefault();
-    subPasswordGet();
-  }
-
-  // 임시 비밀번호 발급 //
-  async function subPasswordGet() {
-    try {
-      const fullEmail = `${form.sendemail}@skuniv.ac.kr`;
-
-      const requestData = {
-        loginId: fullEmail,
-      };
-
-      const response = await APIService.private.post(import.meta.env.VITE_APP_FIND_PASSWORD, requestData);
-
-      // message에서 임시 비밀번호만 얻어서 pw에 저장.
-      const message = response.message;
-      const match = message.match(/임시 비밀번호는\s+(.+?)\s+입니다\./);
-      const pw = match ? match[1] : null;
-
-      // verified가 false인 경우도 처리
-      if (response.success === true) {
-        setSubPassword(pw);
-        setEmail(form.sendemail);
-        setNow(2); // 2번째 페이지 보여줌.
-      } else {
-        alert(response.message);
-      }
-    } catch {
-      // 서버 응답 자체가 실패한 경우
-      alert('서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요.');
-    }
-  }
+  useTimerEmailConfirm(form, setForm, setErrors, count, setCount, setM, setS);
 
   return (
     <div className={styles['passwordFind-form']}>
@@ -224,8 +52,10 @@ export default function PasswordFindForm({ emailSuccess, setEmailSuccess, setEma
                         ? styles['valid']
                         : ''
                 }
-                onChange={inputChange}
-                disabled={sendSuccess === 3 ? true : false}
+                onChange={function (e) {
+                  inputChange(e, setForm, setSendSuccess);
+                }}
+                disabled={sendSuccess === 3 || form.email_valid ? true : false}
                 autoComplete='off'
                 required
               ></input>
@@ -237,7 +67,19 @@ export default function PasswordFindForm({ emailSuccess, setEmailSuccess, setEma
                     ? styles['passwordFind-form__inputbutton']
                     : styles['passwordFind-form__inputbuttonYet']
                 }
-                onClick={handleSendingClick}
+                onClick={function (e) {
+                  handleSendingClick(
+                    e,
+                    form,
+                    setForm,
+                    errors,
+                    setErrors,
+                    setConfirms,
+                    setCount,
+                    setSendSuccess,
+                    handleEmailchecking,
+                  );
+                }}
               >
                 {sendSuccess === 3 ? '전송중' : '인증번호 전송'}
               </button>
@@ -267,7 +109,9 @@ export default function PasswordFindForm({ emailSuccess, setEmailSuccess, setEma
                             ? styles['valid']
                             : ''
                     }
-                    onChange={inputChange}
+                    onChange={function (e) {
+                      inputChange(e, setForm, setConfirmSuccess);
+                    }}
                     disabled={!form.timing}
                     autoComplete='off'
                     required
@@ -279,7 +123,17 @@ export default function PasswordFindForm({ emailSuccess, setEmailSuccess, setEma
                         ? styles['passwordFind-form__inputbuttonYet']
                         : styles['passwordFind-form__inputbutton']
                     }
-                    onClick={handleCheckingClick}
+                    onClick={function (e) {
+                      handleCheckingClick(
+                        e,
+                        form,
+                        setForm,
+                        setErrors,
+                        setConfirms,
+                        setEmailSuccess,
+                        handleConfirmCodechecking,
+                      );
+                    }}
                   >
                     인증번호 확인
                   </button>
@@ -296,7 +150,9 @@ export default function PasswordFindForm({ emailSuccess, setEmailSuccess, setEma
                     id='confirmCode'
                     value={form.confirmCode}
                     className={errors.confirmCode ? styles['invalid'] : form.confirmCode ? styles['valid'] : ''}
-                    onChange={inputChange}
+                    onChange={function (e) {
+                      inputChange(e, setForm, setConfirmSuccess);
+                    }}
                     autoComplete='off'
                     disabled={true}
                   ></input>
@@ -307,7 +163,17 @@ export default function PasswordFindForm({ emailSuccess, setEmailSuccess, setEma
                         ? styles['passwordFind-form__inputbuttonYet']
                         : styles['passwordFind-form__inputbutton']
                     }
-                    onClick={handleCheckingClick}
+                    onClick={function (e) {
+                      handleCheckingClick(
+                        e,
+                        form,
+                        setForm,
+                        setErrors,
+                        setConfirms,
+                        setEmailSuccess,
+                        handleConfirmCodechecking,
+                      );
+                    }}
                   >
                     인증번호 확인
                   </button>
@@ -345,7 +211,10 @@ export default function PasswordFindForm({ emailSuccess, setEmailSuccess, setEma
             <button
               style={{ cursor: 'pointer' }}
               className={styles['passwordFind-form__button--submittingSuccess']}
-              onClick={next}
+              onClick={function (e) {
+                e.preventDefault();
+                subPasswordGet(form, setEmail, setNow, setSubPassword);
+              }}
             >
               비밀번호 찾기
             </button>
